Improvements to scale_parameters.py
Many changes to ScaleParameters with the purpose of making this more
generally useful.

Improve logging in ScaleParameters around node/broker memory
calculations.

In ScaleParameters we have some values which should be synced with
corresponding defaults in Redpanda. Note the requirement to update
the python values when the C++ ones change.

Additionally, we update the node memory calculation when docker is
used.

The calculation of "node memory" was wrong for docker. We subtracted
the estimated reserve from the DEFAULT_MEMORY_MB (2048) but in this
case we pass --memory=DEFAULT_MEMORY_MB, so that is the correct
Redpanda heap value: reserve is not subtracted from it (it _is_
subtracted from total node memory in the case --memory is not passed
at all).

The result was that Redpanda was passed a very small --memory value of
only ~500 MB. Fix it to use the full DEFAULT_MEMORY_MB.

Additionally, we now "use a higher lower clamp": clamp below the docker
memory per node at DEFAULT_MEMORY_MB not 500.

ScaleParameters would use 500 MB for redpanda memory which is way too
low (with 4 shards that's barely over 100 MB/shard or 10x lower than
our supported lower limit).

Use DEFAULT_MEMORY_MB instead (2048 MB) which is what you get when you
don't use scale parameters anyway.

Consider memory for limits

This class calculates the partition limit for scale tests but it
didn't take into account memory when calculating that limit (but it
did when calculating how much memory was required in the cluster).

Adjust the calculation so it is the lower of the per-shard and per-mb
configured values (as it is at runtime in Redpanda).