k/server/test: reduce quota manager bench reqs
Reduce the runtime of the quota manager benchmark by reducing the number
of requests it loops through.

This is because the benchmark was already near the bazel test timeout
and some upcoming changes push it above the timeout. Instead of
increasing the timeout, I chose to reduce the number of iterations since
this 3M requests are plenty for testing anyway.