`storage`: use `segment_closed_exception` in `segment_utils`
In compaction related functions, we throw a `segment_closed_exception`
when encountering a segment that has been closed concurrently during
the process.

However, there is another location in `make_concatenated_segment()` in
which we throw a `std::runtime_error` with a descriptive message of
the segment closed instead.

Change this throw site to use `segment_closed_exception`.