kafka: introduced datalake throttle manager
Introduced a component that is responsible for throttling Iceberg
enabled producers when partition translation backlog target is not met.

The `datalake_throttle_manager` keeps track of all Iceberg related
producers. The producer is marked as iceberg enabled on any shard, then
the manager map reduces the state and throttling is always calculated on
shard 0.

In order to skip the x-shard call for non-iceberg producers and in case
the throttling is not required, every shard local manager can verify if
the throttling is required. Only if the translation backlog are over the
limit the x-shard call is required to verify if the producer is iceberg
enabled one. (shard 0 is the only one that has producer state from all
other shards).

The throttling is calculated based on the time since the translation
error was reported. Currently the max throttle will be applied after 5
minutes if translation problem persists.

Signed-off-by: Michał Maślanka <michal@redpanda.com>