datalake/t: Port partition_probe updates from v1 translator
Also changes the semantics of the lag_metrics ducktape test. In the new
scheduler regime, scheduled translators may sit idle waiting for translatable
records to appear on the input topic. This waiting may cross a commit interval
boundary, in which case commit lag metrics will not be updated until _after_
new records appear, waking up the translator. Only once those records are
processed (and the translator reconciled with the state of the coordinator),
will the commit lag metrics update to reflect the most recent state. In that
sense, commit lag metrics may tend to slightly overstate the case, particularly
in situations where an input topic is quiet for extended periods.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>