datalake: shutdown datalake managers before stopping partitions
Datalake managers may cache the shared pointer to the state machines.
When partitions are stopped the `raft::consensus` pointer is destroyed.
This may lead to a situation in which a coordinator accesses a state
machine with dangling pointer.

This needs a more systematic solution in the state machine itself but in
order to quickly fix the problem and manage the services lifecycle
correctly we simply stop the manager before stopping all partitions.
This way no datalake services will try to access a `raft::consensus`
instance that has already been removed.

Fixes: CORE-8380

Signed-off-by: Michał Maślanka <michal@redpanda.com>