rpc_compiler: make raw_* private and non-virtual
These raw_ methods are not used in a virtual way and do not need to be
public. Perhaps in the past there was some requirement for them to
virtual or this was just an overright.

Making methods non-virtual improves compile time and disk use, since
when a vtable needs to be generated all virtual methods in the class
must be compiled and for these rpc classes that can happen a lot since
all virtual methods are inline (so there is no "key method" which would
allow the compiler to delegate the virtual table generation to a
single TU, see: https://godbolt.org/z/nf94n5Gnd).