rpc_compiler.py: make protocol non-virtual
The generated _protocol object did not need to be virtual and in fact
it was never subclasses except in one test. Remove the virtual qualifier
and also make the class final except for the test RPC services that
need it.

This significantly reduces the on-disk size of some packages after
compilation, e.g., sizes for raft, kafka, cluster directories and
redpanda binary itself:

```
before:

688     bazel-bin/src/v/raft
3300    bazel-bin/src/v/cluster
4312    bazel-bin/src/v/kafka
990     bazel-bin/src/v/redpanda
```

```
630     bazel-bin/src/v/raft
2453    bazel-bin/src/v/cluster
4290    bazel-bin/src/v/kafka
494     bazel-bin/src/v/redpanda/redpanda
```

Note that the size of the cluster directory is smaller by almost 1 GB.