`storage`: fix return type of `verify_read_iobuf`
This was being implicitly converted to `std::error_code`, and a future
commit will attempt to log the error using `to_string(parser_errc)`.

For that reason, the return type and a single call site where the log
line is to be added is amended with the proper type,
`result<iobuf, parser_errc>`, in order to properly log the error.