iceberg/rest_client: protect client requests with gate
Iceberg REST catalog users may expect that it is safe to destroy the
client after its shutdown method returns. This was not the case for rest
client. Underlying http client doesn't use the dispatch gate therefore
calling `http::client::stop()` method does not guarantee that all
requests are finished.

Fixed the issue by using gate in the `iceberg::rest_client` to wait for
requests to be finished before returning from `stop()` method.

The fix in the `http::client` would be preferable but more involving. We
should address that in future.

Fixes: CORE-9549

Signed-off-by: Michał Maślanka <michal@redpanda.com>