kafka/lag_metrics: Fix a race condition
The type of `ktps` is an `input_range`, due to being a `join_view`.

Since it's an input_range, it can only be iterated once.

Holding it by reference, resulted in a race condition as it was
mutated during iteration on each shard.

Fix the race by taking a copy of the view onto each shard. The
additional copy, `copy` is needed as map_reduce cannot take a
mutable lambda.

Signed-off-by: Ben Pope <ben@redpanda.com>