dl/translation: metrics for input throughput
Adds some quick-and-dirty metrics for throughput of incoming Kafka data
meant only for use in debugging.

E.g. this should not be used for usage based billing, given it doesn't
account for commits into Iceberg, doesn't account for errors in the
translation path, and gets reset when partitions are removed or
shutdown.