`storage`: evict readers in `do_compact_adjacent_segments()`
In `do_compact_adjacent_segments()`, we perform a transfer of the
`replacement` segment to the `target` segment while holding the write locks
for each of the segments.

While this seems safe, unfortunately there is one consideration that is
missing- the existing readers in the `readers_cache`.

For reference, in `do_self_compact_segment()`, before swapping the data
file handles, we make sure to call `readers_cache.evict_segment_readers()`.

It seems we were missing this call in `do_compact_adjacent_segments()` before
performing the same operation.

This oversight could be the root cause for a number of issues with readers
created for segments undergoing adjacent compaction.