http/client: 3-args ctor to use default max_idle_time=1s
http::client constructors are somewhat inconsistent: 1-arg and 2-args
ones set max_idle_time to 1s, while the 4-args one accepts max_idle_time
as the 4th argument with a default of 0

Add 3-args constructor that sets it to 1s as well. It is only called in
datalake::coordinator::rest_catalog_factory::create_catalog, and it will
probably benefit from the change.