`config`: reduce `min_cleanable_dirty_ratio` default
The previous default of `0.5` was chosen to be in line with the default
value of `min.cleanable.dirty.ratio` in Kafka.

However, to be more conservative about the change in `redpanda` behavior
from unconditionally compacting every log every `log_compaction_interval_ms`
to _perhaps_ compacting _some_ logs every `log_compaction_interval_ms`,
a lower default value of `0.2` will be at least a little bit closer to
the current status quo, and perhaps a less dramatic change for users.