k/server: improve full disk produce error response
This improves the client-side observability of the scenario when produce
requests are getting rejected due to brokers running out of disk space.

We change the error code to kafka_storage_error conditional on a
sufficiently up to date produce api version, and add a custom error
message to the response.

Note that some older clients that do not know about kafka_storage_error
would handle it as an unknown_error. While kafka_storage_error and
broker_not_available are retriable, unknown_server_error is not. A
retriable error is preferred for easier recovery once disk health
recovers. Hence, we condition the return on supporting at least api
version 4.