kafka: introduced write at offset state machine
Introduced state machine that exposes an interface allowing callers to
replicate batches only if the requested offset matches.

The stm exposes a replicate method that accepts the expected offset and
last log offset. This way it allows the caller not only to replicate
batches but also to recreate gaps in the offset space.

Signed-off-by: Michał Maślanka <michal@redpanda.com>