bazel: support running benchmarks directly with runfiles
Bazel can lookup runfiles using argv0, but we have to pass that in.
Since we don't have that plumbed down to this layer, we'll just resolve
the symlink in the runfiles code. This is test only code anyways (bazel
enforced), so I'm not worried about the IO here for tests/benchmarks as
this should be for setup only.