k/server: Adjust log levels in delete_records
Reduce the "Truncate partition to offset" log message to debug.  At the
current info level, customers that heavily utilize the DeleteRecords API
will experience a flood of these messages.  As there already are metrics
that track DeleteRecords, reducing this to debug does not reduce
observability of the effects of this API.

Also increase the "Possible failed attempted to truncate partition"
message from info to warn as it does represent a possible issue within
Redpanda.

Will backport this change to improve user experience.

Signed-off-by: Michael Boquard <michael@redpanda.com>