datalake: add per-scheduler disk reservation
The per scheduler disk reservation starts off with 0 units. When
requests are made to it for units it uses the disk manager interface
added in the previous commit to request units from core 0 global pool.

This commit also introduces an interface for releasing all unused units
which will be used by core 0 to bring units back to the global pool.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>