test_macros: add fallback macros for non-g/btest
Test macros are used in the test framework to provide a consistent
interface for test cases. The macros are used to define test assertions
in a framework-neutral way. However, they currently always fall back
to the boost macros if IS_GTEST is not defined, which means they are
also compiled the "boost way" - but when not running in a boost test
this silently does nothing.

This is pretty likely to occur in practice since IS_GTEST is only
defined for the TU itself, but fixture and test utilility code is not
compiled with this macro, so when used from a gtest the asserts do not
fire. This also happens for benchmarks which use neither boost nor
gtest.

In this change, we explicitly identify boost tests with IS_BTEST in a
similar way to gtests, and add a fallback to the macros for
non-gtest-non-boost case which uses vassert. This ensures that the test
macros behave "correctly" more or less correctly regardless of the test
framework being used or when no framework is being used.