`storage`: decompress batch during `batch_timequery()`
`ListOffsets` can currently return incorrect results due to the fact
we would not iterate over the records of a compressed batch. We would
instead output the same offset for each record in the batch hit by the
timequery according to the base offset of the batch itself.

To fix this behavior, decompress the batch and iterate over the records
just like we would in the uncompressed case to do the right thing and
return correct results.

This also has the side-effect of making `batch_timequery()` an asynchronous
function.