app: Wire up node_id_override::ignore_existing_node_id
Force the override even if the node already has a node ID in its local kvstore.

This requires forcing the new ID into 'kvs['configuration_invariants']' to
avoid triggering a guard during controller startup.

Note that (like the existing override) this action violates the letter of the
raft protocol and should only be taken with careful discretion.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>