`storage`: add checks for closed `segment`s in sliding window compaction
The list of `segment`s considered for sliding window compaction is
first created and then only later after a number of scheduling points
is the `_segment_rewrite_lock` held.

This means that `segment`s in the window could be removed & closed by
prefix truncation in the meantime. While there are other fail safes in
place to check for a closed `segment` later in the compaction utils, it
is better to check early and avoid any unnecessary I/O.