kafka: Work-around describing user client quotas
This is a work around for the fact that we don't support user based
quotas yet.  The `kafka-configs` tool will error out when attempting to
describe a principal when the describe quotas handler returns
"unsupported_version" when it encounters a request to describe a user
quota.  This is a problem with implementing support for KIP-554, as
`kafka-configs` is a common kafka tool used to manager users.

As the alter path will still error out if asked to add quotas for a
user, this is an acceptable work around to not return anything when
describing quotas for a user.

Signed-off-by: Michael Boquard <michael@redpanda.com>