ct: Update aggregator to use 'extent_meta'
Previously, the aggregator was generating 'dl_placeholder' batches and
returned the 'model::record_batch' instances. These batches was then
replicated using Raft.

The problem with this is that the 'aggregator' and by extension the
'batcher' both have to know about the L0 metadata that metadata layer
uses. If we will have to change the metadata storage format we will have
to change both 'aggregator' and 'batcher'. In order to avoid this
problem the 'aggregator' now uses 'extent_meta' struct. The struct
contains the same information as 'dl_placheolder' but it's not supposed
to be replicated or persisted. Instead, the caller is supposed to
generate the 'dl_placeholder' batches from 'extent_meta' structs.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>