tests/balancer: ignore reconciliation stuck log line
When partition balancer detection of unavailable node is tested the node
may be suspended for a long time. In this situation `controller_backend`
can detect stuck reconciliation as the last reconciliation timestamp
might have been taken before the node was suspended. The test should
still pass in this case.

Added appropriate pattern not to treat stuck reconciliation log line as
a test failure.

Signed-off-by: Michał Maślanka <michal@redpanda.com>