`storage`: use `chunked_vector` in adjacent compact utility functions
Now that we have moved beyond the limit of two segments per adjacent merge
operation, we need to consider the potential size of the container of
`segment`s being constructed and used throughout the adjacent segment utility
functions.

To avoid any potential oversized allocations or failure to allocates, change
use of `std::vector<>` to the non-contiguous `chunked_vector<>` throughout.