tests/cluster_topology: do not account ssh traffic when shaping traffic
Previously all the packets were directed to the queue that did the
traffic shaping. That was incorrect as we are only interested in certain
IPs to be affected by traffic shaping. Fixed the issue by simplifying
the queue chain, now the root priority queue is using non shaped queue
for all ssh traffic.

Signed-off-by: Michał Maślanka <michal@redpanda.com>