`config`: add `log_compaction_merge_max_ranges`
A tunable property that allow control of the maximum number of ranges
which can be processed in a single round of adjacent merge compaction.

A value below 1 effectively disables adjacent merge compaction, a value of
1 has similar behavior to the previous adjacent merge implementation, and
a value of `std::nullopt` imposes no maximum on the number of ranges
which can be combined at once.

The old behavior of adjacent segment compaction (only one pair of segments
combined per round) can now be achieved via setting:
  * `log_compaction_merge_max_segments_per_range = 2`
  * `log_compaction_merge_max_ranges = 1`