tests/Dockerfile: improve image build
Split main apt packages list into two base-tools, and tool-pkg, with
the former being tools needed for the remainder of the image build, and
the latter being the (more numerous) tools needed to run DT but not for
the image build. This lets us get to the fan-out part of the build
more quickly and reduce the critical path.

Call protobuf install explicitly instead of inside java-dev-tools, for
faster rebuilds when protobuf is unchanged (this changes very rarely).

This shaves a minute or so off the DT build for me.