tests/docker: debug-cache writes to container
debug-cache used to echo the contents of all cache dirs to stdout
during the image build (only when building the debug layer). However,
docker cuts off output at 200KIB, which we hit now that introduced
new caches.

Instead write the contents to a file in the container, which can be
inspected later if needed.