http: Set keepalive for http
The http client can potentially deadlock waiting for data from the
connection. If for some reason the connection was dropped and the RST
packet wasn't propagated the `recv_some` call in the http client will be
blocked for arbitrary amount of time (in some cases the middleware could
detect this and send the RST packet later).

If the keepalive is enabled we should send the proble after the idle
time (set to 5m by this commit). The defaults are chosen to be slightly
higher compared to our defaults for Kafka.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>