rptest: self-test for decode_backtraces
Add a self-test to verify that the decode_backtraces
function in RedpandaService works correctly. It sends a segv fault to
RP, then decodes the backtrace and checks that it contains some
expected text.

We also add a raise_on_failure mode to decode_backtraces which we
enable only in this test so we get the underlying exception if the
backtrace methods fails (normally we just log and move on).