rptest/cluster_health_overview: check health overview eventual & stable
This test was flaky as it takes time for 3-fold replicated
partitions to elect a new leader. At times a few seconds due to the
priority system.

Make rf=3 partitions leadership condition part of the waited check
instead of an assert after a node got down. Check it 5 times (10 sec) to
make sure it's not a stale leadership. Check the next condition multiple
times for the same reason.