tests: deflake partition_reassignments_test.py
test_add_partitions_with_inprogress_reassignments attempts to add a
partition to a topic for which all partitions are undergoing
reassignment.

It accomplished this by issuing a reassignment to all paritions
manually, and checking the success of these commands. These commands
could fail if the autobalancer starts a reassignment for the sake of
balancing the cluster, because a reassignment would already be in
progress.

This change disables the autobalancer for this test such that the manual
reassignment command will not race the autobalancer.

This also allows the test to succeed if adding a partition succeeds, as
the failure behavior was written to keep parity with Kafka. This may be
changed in the future to simply succeed.