`storage`: set `offset`s in `segment_appender::close()`
Here, we perform a flush & truncation of the underlying file, but were not
updating the corresponding offsets in the `segment_appender`.

This means that a closed `appender` will be reporting the untruncated/flushed
offsets if queried.

Fix the book-keeping by setting the `_fallocation_offset`, `_flushed_offset` and
`_stable_offset` after truncating in `close()`.