kafka/server: reduce read concurrency in `fetch_ntps_in_parallel`
Reading all ntps in parallel results in the need to reserve memory space
for every ntp all at once. This causes issues when trying to fetch many
partitions from a given shard. In this case only the first N partitions
are likely to get memory units. And of those if few to none have data to
read then little to no data will be returned. The end result of this is
for wide fetches with many partitions Redpanda will often have smaller
than expected responses.

By reducing concurrency in ntp reads the issue is avoided. However,
there is some concerns how this would affect performance in general. To
that ends a given ntp fetch can largely be spilt into three categories;
a read from cloud storage, a read from local storage, or a read from the
batch cache. When all ntps are read from the batch cache the change will
have no effect on performance. And for cases when it is being read from
the disk or cloud storage, data has shown in the past that these reads
largely entail historical reads that are likely to be large per-ntp and
therefore wouldn't require parallelism to saturate the disk.