datalake/translator: add trace logging for flaky debugging
This adds some trace logging around Iceberg translation. There's a flaky
test where a parquet file was written containing a single row after the
lag threshold was reached, and then the rest of the records were written
to a second file immediately after. It's not clear from the logs how
this happened, so I'm adding a bit of trace logging that might help
locate the problem if it reoccurs.