ct/l1: add initial add_objects state_update
Adds an abstraction to represent safely (e.g. no gaps, no overlap) add a
group of L1 objects to the `state`. This abstraction checks that
partition offset ranges are valid before allowing the operation to be
applied.

In subsequent work, the update abstraction will be used as replicated
commands for a replicated STM. As such, they are serde-able.