ct/l1: add simple_metastore
Adds an implementation of the metastore interface that is backed by the
newly introduced `state`. This metastore is not replicated or persisted
at all, but the state_updates that will be used for the STM are used
here as well in, so there should be little to no behavioral differences
to the eventual replicated-STM-backed metastore. Since metastore only
depends on simple struct definitions, it is ideal for parallelizing work
that doesn't require end-to-end testing.