test: deflake kafka time query test
When the test starts up retention is scheduled inside of Kafka at
30seconds, and then periodically after that every 5 minutes.

    [2025-06-23 18:42:50,481] DEBUG Scheduling task kafka-log-retention with initial delay 30000 ms and period 300000 ms. (kafka.utils.KafkaScheduler)

The first thing the test does is creates a topic and starts writing data
into it.

    [DEBUG - 2025-06-23 18:43:07,811 - kafka_cli_tools - _execute - lineno:675]: Created topic tqtopic.

    [DEBUG - 2025-06-23 18:43:12,015 - kgo_verifier_services -
    _ingest_status - lineno:428]: KgoVerifierProducer-0-281471707496992
    status: [{'topic': 'tqtopic', 'sent': 1025, 'acked': 0,

While the produce is going on, the initial retention check in Kafka runs
on all three brokers, with nothing to do because the default retention
size and time is much larger than what has been produced.

    [2025-06-23 18:43:20,084] DEBUG Garbage collecting 'tqtopic-0' (kafka.log.LogManager)
    [2025-06-23 18:43:20,086] DEBUG Log cleanup completed. 0 files deleted in 0 seconds (kafka.log.LogManager)

    [2025-06-23 18:43:19,802] DEBUG Garbage collecting 'tqtopic-0' (kafka.log.LogManager)
    [2025-06-23 18:43:19,803] DEBUG Log cleanup completed. 0 files deleted in 0 seconds (kafka.log.LogManager)

    [2025-06-23 18:43:20,484] DEBUG Garbage collecting 'tqtopic-0' (kafka.log.LogManager)
    [2025-06-23 18:43:20,485] DEBUG Scheduling task kafka-delete-logs with initial delay 60000 ms and period -1 ms. (kafka.utils.KafkaScheduler)

After this, produce ends

    [DEBUG - 2025-06-23 18:43:22,026 - kgo_verifier_services -
    _ingest_status - lineno:428]: KgoVerifierProducer-0-281471707496992
    status: [{'topic': 'tqtopic', 'sent': 3072, 'acked': 3072,

And finally retention is set to be small (for the test) in order to
trigger the removal of data and alter the starting offset

    [DEBUG - 2025-06-23 18:43:22,589 - rpk - _execute - lineno:1217]:
    Executing command:
    ['/var/lib/buildkite-agent/builds/arm64-xfs-builders-i-02d050a5b241dfbc6-1/redpanda/redpanda/vbuild/redpanda_installs/ci/bin/rpk',
    'topic', '-X',
    'brokers=docker-rp-15:9092,docker-rp-11:9092,docker-rp-23:9092',
    'alter-config', 'tqtopic', '--set', 'retention.bytes=1048576',

The test then proceeds to wait 120 seconds for retention to be applied
and the starting offset to change. However, the it will need to wait up
to 5 minutes for this to happen because it missed the initial retention
pass that happened 30 seconds after start up.

This commit addresses the issue by configuring Kafka to run retention
more frequently (every 30 seconds) so that it will run within the 120
second window given.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>