rpk: fix concurrent map writes in proto serde encodefn
When serde.EncodeRecord is called on a cached serde function across
multiple goroutines, dynamicpb.Set panics with a concurrent map writes
fatal. This is caused by the message previously being created outside
of the returned encodefn and attempting to unmarshal the byte slice
on the same dynamicpb message.

While not an issue for rpk per-se, if this package is imported into
another project which does attempt concurrent encodes, this fails.
By moving the creation of the dynamicpb object inside the returned
encodefn, it does not impact the performance of rpk but allows
concurrent encoding to happen if other packages chose to import it.