`storage`: add `ntp_config::tx_retention_ms()`
This commit adds an accessor function to `ntp_config` for
`tx_retention_ms`.

It also slightly reworks the logic in `ntp_config::tombstone_retention_ms()`
to make use of a new, generic `ntp_config::delete_retention_ms()`.

The checks in `tombstone_retention_ms()` and `tx_retention_ms()` differ
because we need to ensure that the tombstone retention time is always
`std::nullopt` when tiered storage is enabled (as there still exists
race conditions with tombstone removal and tiered storage uploads),
whereas `tx_retention_ms()` does not have the same restriction (we always
upload below LSO, and the removal of local transactional batches does not
impact the correctness of the log if the same transactional batches
still exist in the cloud).