tx/tm_stm: remove dropped_pid based cleanup
This is a very old code but the intent seems to be that it tracks
previous epoch pids so they are cleaned up. But the code in the current
form is just wrong the way it is structured, the cleanup happens only on
the leader but not followers. Instead the next commit consolidates
cleanup of older epochs in apply so its done on all replicas