`cloud_storage_clients`: handle explicit backend in `s3_client::self_configure()`
If a user manually set the `cloud_storage_backend`, it should be considered
the source of truth for self configuration.

Check its value and use the same `path` override for `oracle_s3_compat`
and `minio` in this early return path before checking inferred backend
or performing a manual `ListObjects` request to self-configure.