rptest: stop redpanda on failure path
In the @cluster decorator after-test path, we stop the redpanda cluster
asssociated with test in the success path but not the failure path.

This causes issues when decoding the backtrace since the decoding may
have a large in-memory working set which competes with Redpanda itself
which can still be running and consuming almost all memory.

This can slow down the decoding process a lot. So stop Redpanda first,
and this should be safe as we already do it on the happy path.

Fixes CORE-12786.