iceberg/config: validate REST endpoint configuration
This disarms a footgun. When Iceberg is enabled, but no REST catalog
is set, Redpanda fails to start because of the invalid
configuration. This adds validation of the property pair that applies
when config is altered via the API, so users are not surprised when
they alter the config and Redpanda fails to restart.

This is a stopgap measure. The longer term goal is to allow users to
configure Iceberg properties without requiring a restart.