dt/si: Dial timeouts way down in fast uploads mode
In at least one ducktape test, we can get into a situation where a leadership
transfer followed by a faulty volley of segment uploads can eat away most of
our time budget for the test. In fast_upload mode, we expect segments on the
smaller side, and it's convenient to let them fail quickly back to be retried.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>