kc/verifier: added verifier application for direct consumer
Added simple test application that will allow basic scale testing of a
direct consumer. The application exposes REST API that the user can
interact with to drive the direct consumers to fetch from the topic.

So far the application doesn't have any way of consuming the fetched
data but it returns per partition statistics that can be used to
validate if the consumer is making progress.

Signed-off-by: Michał Maślanka <michal@redpanda.com>