rptest: Remove outdated recovery tests
The topic_recovery_test has a bunch of test cases which are checking the
deep recovery. These tests are written in a way which will become
problematic when we will move to using log_reader in the ntp_archiver.
These tests are removed now. Few newer tets are still there but they
have to be updated to stay useful.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>