ct: consolidate level zero infrastructure bits
The cloud_topics/core directory contained the pipeline bits which are
currently used only by level zero. These bits are written to be generic,
and the thought is that we will use them in other parts of cloud_topics.
However, we want to accept some churn in the tree in order to normalize
the layout, and pull out the common bits as they are identified. The
same goes for the throttler as well.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>